Clarify the registration limit
While initially reading the manifesto, I understood it as if the number
of registration should be limited per IP per hour, i.e. the only limit
is when you make too many registrations per hour from a particular IP,
then registrations from that IP get blocked for an hour.

However, I've seen someone else unrestand the previous wording
as if there were to be two separate limits, one per hour
regardless of IP, and one per IP for the entire server lifetime.

The latter doesn't make much sense.

Change the wording so that it's more clear that it's about the
first meaning.